+++ test/multiCompiler.test.js
@@ -49,7 +49,8 @@ describe('Plugin', function () {
         two: {
           js: 'two-bundle.js'
         }
-      }
+      },
+      images: []
     }
 
     var args = {
@@ -82,7 +83,7 @@ describe('Plugin', function () {
       plugins: [new Plugin({path: 'tmp', update: true})]
     }
 
-    var expected = {chunks: {one: {js: 'one-bundle.js'}, two: {js: 'two-bundle.js'}}}
+    var expected = {chunks: {one: {js: 'one-bundle.js'}, two: {js: 'two-bundle.js'}}, images: []}
     var args = {config: config_2, expected: expected}
 
     webpack(config_1, function (err, stats) {
@@ -114,7 +115,7 @@ describe('Plugin', function () {
       plugins: [new Plugin({path: 'tmp', update: false})]
     }
 
-    var expected = {chunks: {two: {js: 'two-bundle.js'}}}
+    var expected = {chunks: {two: {js: 'two-bundle.js'}}, images: []}
     var args = {config: config_2, expected: expected}
 
     webpack(config_1, function (err, stats) {