+++ CHANGELOG.md
@@ -10,6 +10,16 @@ This change log follows the format documented in [Keep a CHANGELOG].
 
 ## Unreleased
 
+## 4.0.0
+
+### Changed
+
+- Assets for entry points are now located in `chunks` object
+
+### Added
+
+- Images are extracted and saved in `images` array
+
 ## 3.4.0 - 2016-03-09
 
 ### Changed